SMR-22 feat: change stockDisponible type from BigDecimal to Integer in UpdateProductDTO

+++ src/main/java/com/smartshop/dto/requist/updateRequistDto/UpdateProductDTO.java
@@ -23,6 +23,6 @@ public class UpdateProductDTO {
 
     @DecimalMin(value = "0", message = "Stock must be 0 or greater")
     @Digits(integer = 10, fraction = 2, message = "Stock must have maximum 10 integer digits and 2 decimal digits")
-        private BigDecimal stockDisponible;
+        private Integer stockDisponible;
 }
